pyproject.toml
[project]
name = "dss-provisioner"
dynamic = ["version"]
description = "Terraform-style infrastructure-as-code management for Dataiku DSS"
readme = "README.md"
authors = [
    { name = "Jonas Meyer-Ohle", email = "19151471+jonas-meyer@users.noreply.github.com" }
]
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.10",
    "pydantic-settings>=2.7",
    "dataiku-api-client>=14.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8",
    "ty>=0.0.1a7",
    "pytest>=8.0",
    "pytest-cov>=6.0",
    "testcontainers>=4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/dss_provisioner"]

[tool.hatch.version]
path = "src/dss_provisioner/__init__.py"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "PTH",    # flake8-use-pathlib
    "RUF",    # ruff-specific rules
]
ignore = [
    "E501",   # line too long (handled by formatter)
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ty.environment]
python-version = "3.13"